[MG Framework] Refactor Mouse
