+++ MonoGame.Framework/Input/Mouse.Default.cs
@@ -1,0 +1,32 @@
+// MonoGame - Copyright (C) The MonoGame Team
+// This file is subject to the terms and conditions defined in
+// file 'LICENSE.txt', which is part of this source code package.
+
+using System;
+
+namespace Microsoft.Xna.Framework.Input
+{
+    public static partial class Mouse
+    {
+        private static IntPtr PlatformGetHandle()
+        {
+            return IntPtr.Zero;
+        }
+
+        private static MouseState PlatformGetState(GameWindow window)
+        {
+            return window.MouseState;
+        }
+
+        private static void PlatformSetPosition(int x, int y)
+        {
+            PrimaryWindow.MouseState.X = x;
+            PrimaryWindow.MouseState.Y = y;
+        }
+
+        public static void PlatformSetCursor(MouseCursor cursor)
+        {
+
+        }
+    }
+}
